Functions seperated into two different files for testing purposes
